Add specific nat rules for inbound envoy routing
Istio Envoy treats differently inbound traffic that
targets ports declared in a service and ports
that are undeclared in a service.

Traffic that targets ports undeclared in an associated
service is sent from 127.0.0.6 to <POD_IP>, instead of
being sent from localhost to localhost.

This commit adds rules that hit traffic sent from
127.0.0.6 and dnats it to the VM

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>